Pizza Monitor/services/detection/requirements.txt
ultralytics>=8.3.0
torch>=2.9.0 --index-url https://download.pytorch.org/whl/cpu
torchvision>=0.20.0 --index-url https://download.pytorch.org/whl/cpu
opencv-python-headless>=4.8.0
pika>=1.3.2
psycopg2-binary>=2.9.9
numpy>=1.24.3,<2  